requirements.txt
anthropic>=0.96.0
python-telegram-bot==21.6
notion-client==2.2.1
APScheduler==3.10.4
newsapi-python==0.2.7
feedparser==6.0.11
requests==2.31.0
python-dotenv==1.0.1
openai>=1.12.0
pytz==2024.1
